refactor(v1): split ecosystem configs into per-ecosystem files
configs.tsx held both collectorConfig and javaAgentConfig in one 250-line file. Move each into its own collector-config.tsx / java-agent-config.tsx (each carrying only the imports and path constant it uses) and import them directly from the matching landing wrapper. Adding an ecosystem is now a new sibling config file, not an edit to a shared blob.
